ml/gnn_lexicon/requirements.txt
# Philippine Lexicon GNN Requirements

# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0

# Database
psycopg2-binary>=2.9.0

# Scientific computing
numpy>=1.21.0
scikit-learn>=1.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
black>=22.0.0
flake8>=4.0.0
mypy>=0.990 